requirements.txt
Flask
OpenAI
python-dotenv
